backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary>=2.9.9
alembic==1.13.2

# Data validation
pydantic[email]==2.9.2
pydantic-settings>=2.5.2
python-dotenv==1.0.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Real-time
redis>=5.1.0

# AI / LLM
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.1.1
langchain-huggingface>=0.1.0 

# Market data
httpx==0.27.2
pandas>=2.2.2
numpy>=2.0.0

# PDF extraction
pypdf>=3.0.0

# Vector store
faiss-cpu>=1.7.4